# web/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "API интерфейс SAMPLE",
...
        "contact": {
            "name": "Nikolaj Rudakov",
            "email": "nnrudakov@gmail.com",
            "url": "https://github.com/nnrudakov/sample-api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.sample.nr",
            "description": "Тестовый сервер"
        }
    ],
    "paths": {
        "/companies/{id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получение информации об организации",
                "description": "Возвращает информацию об организации по её идентификатору.\n\nАтрибуты по типам ([см. использование](#section/Poluchenie-atributov-sushnostej)):\n\n| Атрибут        | По умолчанию  | Расширенные  |\n|----------------|:-------------:|:------------:|\n| id             | *             |              |\n| title          | *             |              |\n| inn            | *             |              |\n| ogrn           |               | *            |\n| address        | *             |              |\n| phone          | *             |              |\n| person         | *             |              |\n| contact        | *             |              |\n| enabled        | *             |              |\n| comment        |               | *            |",
                "operationId": "companies-get-one",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Выбор атрибутов по типам",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Attributes"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyFull"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Организации"
                ],
                "summary": "Обновление организации",
                "description": "Получает новые данные организации и обновляет.",
                "operationId": "companies-update",
                "parameters": [
                    {
                        "name": "X-Csrf-Token",
                        "in": "header",
                        "description": "CSRF токен",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Новые данные",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "$ref": "#/components/schemas/title"
                                    },
                                    "inn": {
                                        "$ref": "#/components/schemas/inn"
                                    },
                                    "ogrn": {
                                        "$ref": "#/components/schemas/ogrn"
                                    },
                                    "address": {
                                        "$ref": "#/components/schemas/address"
                                    },
                                    "phone": {
                                        "$ref": "#/components/schemas/phone"
                                    },
                                    "person": {
                                        "$ref": "#/components/schemas/person"
                                    },
                                    "contact": {
                                        "$ref": "#/components/schemas/contact"
                                    },
                                    "enabled": {
                                        "title": "Активна",
                                        "type": "boolean"
                                    },
                                    "comment": {
                                        "$ref": "#/components/schemas/comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Данные успешно изменены"
                    },
                    "400": {
                        "description": "Неверный CSRF токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибки валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidateError"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "Ошибка",
                "properties": {
                    "name": {
                        "title": "Имя ошибки",
                        "description": "Имя ошибки может использоваться в заголовках диалоговых окон.",
                        "type": "string"
                    },
                    "message": {
                        "title": "Текст ошибки",
                        "description": "Текст ошибки для пользователя.",
                        "type": "string"
                    },
                    "code": {
                        "title": "Код ошибки",
                        "description": "Код ошибки внутри системы.<br>0 — Неизвестная ошибка;<br>1 — Объект не найден;<br>2 — Неверный CSRF токен;<br>3 — Неверные параметры запроса;<br>4 — Ошибка входа/выхода в систему;<br>5 — Доступ запрещён",
                        "type": "integer",
                        "format": "int32",
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5
                        ]
                    },
                    "status": {
                        "title": "HTTP статус",
                        "description": "[HTTP](https://en.wikipedia.org/wiki/List_of_HTTP_status_codes) статус ошибки.",
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "title": "Исключение сервера",
                        "description": "Исключение сервера, которое вызвало ошибку. Присутствует только в режиме разработки.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidateError": {
                "title": "Ошибки валидации",
                "properties": {
                    "field": {
                        "title": "Имя поля",
                        "description": "Имя поля, в котором произошла ошибка.",
                        "type": "string"
                    },
                    "message": {
                        "title": "Текст ошибки",
                        "description": "Текст ошибки для пользователя.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Attributes": {
                "title": "Выбор атрибутов по типам",
                "properties": {
                    "fields": {
                        "title": "Атрибуты по умолчанию",
                        "description": "Список атрибутов через запятую.",
                        "type": "string"
                    },
                    "expand": {
                        "title": "Расширенные атрибуты",
                        "description": "Список атрибутов через запятую.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "id": {
                "title": "Идентификатор",
                "type": "integer",
                "format": "int64"
            },
            "title": {
                "title": "Наименование",
                "type": "string"
            },
            "comment": {
                "title": "Комментарий",
                "type": "string"
            },
...
            "inn": {
                "title": "ИНН",
                "type": "string"
            },
            "ogrn": {
                "title": "ОГРН",
                "type": "string"
            },
            "address": {
                "title": "Адрес",
                "type": "string"
            },
            "phone": {
                "title": "Телефон",
                "description": "Формат номера `+7XXXXXXXXXX`.",
                "type": "string"
            },
            "person": {
                "title": "Ответственное лицо",
                "description": "ФИО ответственного лица.",
                "type": "string"
            },
            "contact": {
                "title": "Контакт ответственного лица",
                "type": "string"
            },
            "enabled": {
                "title": "Активность",
                "description": "Является ли организация не заблокированной.",
                "type": "boolean"
            },
...
            "CompanyFull": {
                "title": "Организация",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "title": {
                        "$ref": "#/components/schemas/title"
                    },
                    "inn": {
                        "$ref": "#/components/schemas/inn"
                    },
                    "ogrn": {
                        "$ref": "#/components/schemas/ogrn"
                    },
                    "address": {
                        "$ref": "#/components/schemas/address"
                    },
                    "phone": {
                        "$ref": "#/components/schemas/phone"
                    },
                    "person": {
                        "$ref": "#/components/schemas/person"
                    },
                    "contact": {
                        "$ref": "#/components/schemas/contact"
                    },
                    "enabled": {
                        "$ref": "#/components/schemas/enabled"
                    },
                    "comment": {
                        "$ref": "#/components/schemas/comment"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}